fix:: Reorganise the file/module structure to fix warnings
The compiler was reporting a lot of unused functions in the model module
which turned out to be the problem with me not understanding the proper
module structure in Rust.

The solution is that all modules (except main) plugs into the lib.rs
file, which then able to serve both the integration tests and the main
module.

Effectively, this separates the functionality into a the "banking" crate
and the entry point for the application (banking).
